Implement non-blocking UDB search and make UDB registration on login
Registration for user discovery will now be part of the registration flow.
On the first login, user email can be passed to login function, which will
call the internal registerForUserDiscovery API.

This way, the only UDB command that can be run from command line now is
a SEARCH, since the REGISTER is already done.